+++ CHANGELOG.md
@@ -10,6 +10,8 @@ project adheres to [Semantic Versioning][SemVer].
 ### Fixed
 - Don't crash without explicit `listen` configuration. This bug was introduced
   with version 1.7.0.
+- Don't crash if the configuration file is empty (i.e., has no `eturnal`
+  section).
 
 ## [1.8.1] - 2022-01-10
 ### Fixed